+++ src/EventEmitter.js
@@ -3,13 +3,19 @@ import * as utils from './utils/utils';
 
 export default class EventEmitter extends EventEmitter2 {
   constructor(conf = {}) {
-    const { loadLimit = 50, eventsSafeInterval = 300, pause = 500, ...ee2conf } = conf;
+    const {
+      loadLimit = 50,
+      eventsSafeInterval = 300,
+      pause = 500,
+      inspect = false,
+      ...ee2conf
+    } = conf;
     super(ee2conf);
 
     const [isPaused, togglePause] = utils.withSwitch(false, true);
 
     const overloadHandler = () => {
-      console.warn('Infinite loop detected', this.id, pause);
+      console.warn('Infinite loop detected');
       togglePause();
       setTimeout(togglePause, pause);
     };
@@ -20,6 +26,10 @@ export default class EventEmitter extends EventEmitter2 {
     });
 
     this.emit = (...args) => {
+      if (typeof inspect === 'function') {
+        inspect();
+      }
+
       if (isPaused()) {
         return;
       }