+++ src/components/datagrid/DataGrid.js
@@ -37,6 +37,9 @@ export default class DataGridComponent extends NestedComponent {
       const groups = _.get(this.component, 'rowGroups', []);
       const rowsNum = this.totalRowsNumber(groups);
       this.setStaticValue(rowsNum);
+      this.dataValue = _.zipWith(this.dataValue, this.defaultValue, (a, b) => {
+        return _.merge(a, b);
+      });
     }
   }