+++ test/forms/index.js
@@ -8,6 +8,9 @@ import SubmissionSetter from './submissionSetter';
 import NestedFormValidation from './nested-form-validation';
 import NestedFormTests from './nested-form-tests';
 import NestedFormNoSubmit from './nested-nosubmit.js';
+import NestedFormEventBubbling from './nested-event-bubbling';
+import NestedConditionallyHidden from './conditional-nested-form-load.js';
+
 import ChildMetadata from './childMetadata';
 
 export default [
@@ -21,5 +24,6 @@ export default [
   NestedFormValidation,
   NestedFormTests,
   NestedFormNoSubmit,
+  NestedConditionallyHidden,
   ChildMetadata
 ];