+++ test/wizards/index.js
@@ -1,3 +1,4 @@
+import './infinite-loop.test';
 import Simple from './simple';
 import Conditional from './condional.next.page';
 export default [